Refactor: business 데이터 파일 타입 분리 및 정리 - 모든 interface 정의를 types/components/business.ts로 이동 - 데이터 파일들에서 타입 정의 제거, import type 사용 - index.ts에서 타입 export 제거, 데이터만 export - MenuItem, ProductItem에 image 속성 추가 - 105줄 삭제, 56줄 추가로 코드 간소화 - 타입과 데이터의 명확한 분리로 유지보수성 향상
