+++ src/components/business/data/index.ts
@@ -1,13 +1,8 @@
 // Urban Regeneration Data
-export {
-  academyCards,
-  educationCards,
-  type AcademyCard,
-  type EducationCard,
-} from './urbanRegenerationData';
+export { academyCards, educationCards } from './urbanRegenerationData';
 
 // Events Education Data
-export { eventCards, type EventCard } from './eventsEducationData';
+export { eventCards } from './eventsEducationData';
 
 // Cafe27b Data
 export * from './cafe27bData';
